Min price update moved from the thread to one-time call in audit.py
