+++ src/qsp_protocol_node/audit/threads/__init__.py
@@ -5,7 +5,6 @@
 #                                                                                                  #
 ####################################################################################################
 
-from .update_min_price_thread import UpdateMinPriceThread
 from .claim_rewards_thread import ClaimRewardsThread
 from .compute_gas_price_thread import ComputeGasPriceThread
 from .collect_metrics_thread import CollectMetricsThread
@@ -16,7 +15,6 @@ from .submit_report_thread import SubmitReportThread
 from .poll_requests_thread import PollRequestsThread
 
 __all__ = ['QSPThread',
-           'UpdateMinPriceThread',
            'ClaimRewardsThread',
            'ComputeGasPriceThread',
            'CollectMetricsThread',
